nlp/requirements.txt
llama-cpp-python
fire